# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "API de Líneas de Microbuses",
        "version": "1.0.0",
        "description": "Documentación de líneas, rutas y puntos (ida/vuelta)."
    },
    "paths": {
...
        "/api/rutas/planificar/": {
            "post": {
                "operationId": "planear_ruta_por_coordenadas",
                "description": "Recibe coordenadas de inicio y fin, construye un grafo con:\n- Tramos de bus (LineaRuta).\n- Caminatas desde el origen a paradas cercanas.\n- Caminatas entre paradas cercanas (trasbordos caminando).\n- Caminatas desde paradas cercanas hasta el destino.\n\nUtiliza Dijkstra con el tiempo como peso, penaliza las caminatas (tiempo más alto que el bus), penaliza los trasbordos y limita a 2 cambios de línea.",
                "summary": "Ruta más corta con máximo 2 trasbordos (considerando caminata)",
                "tags": [
                    "rutas"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RutaPlanRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/RutaPlanRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/RutaPlanRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "basicAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RutaPlanificada"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Coordenadas": {
                "type": "object",
                "properties": {
                    "lat": {
                        "type": "number",
                        "format": "double"
                    },
                    "lon": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "lat",
                    "lon"
                ]
            },
...
            "LineaTrayecto": {
                "type": "object",
                "properties": {
                    "codigo": {
                        "type": "string"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "sentido": {
                        "type": "integer"
                    }
                },
                "required": [
                    "codigo",
                    "nombre",
                    "sentido"
                ]
            },
...
            "PuntoRutaPlanificada": {
                "type": "object",
                "properties": {
                    "orden": {
                        "type": "integer"
                    },
                    "latitud": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,3}(?:\\.\\d{0,6})?$"
                    },
                    "longitud": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,3}(?:\\.\\d{0,6})?$"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "distancia_tramo": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "distancia_acumulada": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "tiempo_tramo": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$"
                    },
                    "tiempo_acumulado": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$"
                    },
                    "linea_trasbordo": {
                        "type": "string",
                        "nullable": true,
                        "description": "Código de la línea a la que se hace trasbordo en este punto (o null si no hay trasbordo)."
                    }
                },
                "required": [
                    "descripcion",
                    "distancia_acumulada",
                    "distancia_tramo",
                    "latitud",
                    "linea_trasbordo",
                    "longitud",
                    "orden",
                    "tiempo_acumulado",
                    "tiempo_tramo"
                ]
            },
...
            "RutaPlanRequest": {
                "type": "object",
                "properties": {
                    "inicio": {
                        "$ref": "#/components/schemas/Coordenadas"
                    },
                    "fin": {
                        "$ref": "#/components/schemas/Coordenadas"
                    }
                },
                "required": [
                    "fin",
                    "inicio"
                ]
            },
            "RutaPlanificada": {
                "type": "object",
                "properties": {
                    "lineas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LineaTrayecto"
                        }
                    },
                    "descripcion_ruta": {
                        "type": "string"
                    },
                    "distancia_total": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "tiempo_total": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,4}(?:\\.\\d{0,2})?$"
                    },
                    "puntos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PuntoRutaPlanificada"
                        }
                    }
                },
                "required": [
                    "descripcion_ruta",
                    "distancia_total",
                    "lineas",
                    "puntos",
                    "tiempo_total"
                ]
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}